refactor(refactor): rename runtime ns to core
